Fix Linux issues
On Linux (at least Ubuntu 16.04), m_portPath does not have a trailing separator character, we we need to add one. QDir::cleanPath() will remove duplicate separators, for example on Windows.

Also, on Linux, files are not commited to disk immediatly, so we need to call sync before restarting after updating the color map.